Updates the code that reads files to take into account a list of excluded glob patterns
