+++ src/core/markdown/adapters/__tests__/interface-adapter.spec.ts
@@ -17,6 +17,7 @@ const defaultMarkdownGeneratorConfig: MarkdownGeneratorConfig = {
   sortAlphabetically: false,
   linkingStrategy: 'relative',
   referenceGuideTitle: 'Apex Reference Guide',
+  exclude: [],
   excludeTags: [],
 };
